requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
emoji==2.2.0
idna==3.4
names==0.3.0
oauthlib==3.2.2
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
tweepy==4.12.1
urllib3==1.26.14
